requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

sqlalchemy==2.0.35
asyncpg==0.29.0                # Async PostgreSQL driver
psycopg2-binary==2.9.9         # Sync PostgreSQL driver
alembic==1.13.3                # Database migrations
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.2.0
redis==5.2.0
hiredis==3.0.0
celery==5.4.0
flower==2.0.1                  # Celery monitoring

# Selenium for automation
selenium==4.26.1
webdriver-manager==4.0.2      
pillow==11.0.0                 # For screenshot processing

# Excel & Google Sheets parsing
openpyxl==3.1.5                # Excel files
pandas==2.2.3                  # Data processing
gspread==6.1.4                 # Google Sheets API wrapper
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# HTTP Client
httpx==0.27.2
aiohttp==3.11.7

# Utilities
python-dateutil==2.9.0
pytz==2024.2
pydantic-extra-types==2.10.1

# Logging & Monitoring
loguru==0.7.2
sentry-sdk[fastapi]==2.18.0    # Error tracking

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2                  # For testing FastAPI
faker==33.0.0                  # Fake data generation

# Development
black==24.10.0                 # Code formatter
isort==5.13.2                  # Import sorter
flake8==7.1.1                  # Linter
mypy==1.13.0                   # Type checker
pre-commit==4.0.1              # Git hooks

# Documentation
mkdocs==1.6.1
mkdocs-material==9.5.47

# Environment variables
python-dotenv==1.0.1
